+++ CHANGELOG.md
@@ -19,3 +19,11 @@
 ### 2026-05-01 Fixed
 
 - Replace 'ns0:' with empty string in XML parsing to ensure compatibility with different ONIX file formats.
+
+## [0.1.2] - 2026-05-02
+
+### 2026-05-02 Added
+
+- Added `authors` property to `Product` model to retrieve all contributors with role "A01" (written by) and "A02" (co-written by).
+- Added `editor` property to `Product` model to retrieve gln and name of the first publisher in supplier detail, if available.
+- Added `price` property to `Product` model to retrieve the first price of the product, if available (ttc, ht, currency, vat rate).